Move style tags to top of shell and sections
When I deployed to production in #22 I noticed a flash
of context when the page loaded and I believe that the
reason is that it takes longer to load the larger list
of branches than it does on my local machine, and when
the stylesheets are all at the end they don't yet load
until the rest of the page has as well.

In this patch I'm moving styles to the top in hopes
that the browser will parse those immediately and
eliminate the flash of unstyled content.