Report uncaught exceptions
Tracks and reports uncaught exceptions and unhandled promise rejections.
Also adds a listing of unhandled promise rejections to the debug screen.
Unhandled promise rejections can linger around and come and go so in
this case we don't just report them as they come but rather wait for a
bit and see if they linger.