DO NOT MERGE: Adds ability to remove images and containers
**DO NOT MERGE**

This is a work-in-progress patch to allow the manual removal of any
image or container on the running server from the `/debug` page.

This _would_ be working except it introduces a bug: when we delete
by container or image id we leave behind the build directory for the
commit hash the container or image was originally associated with. This
means that whenever someone tries to load this commit hash again
then `dserve` will thing the container is being build.

We need to associate build dirs not only with commit hash but also
with the images they build or make sure to delete them after the
images themselves are finished building.