Initial implementation of middleware for subdomains.
Current versions of dserve use a single domain for all commits and
branches.  This can allow cross-contamination of client-side resources
like cookies, localStorage and assets cached by service workers across
different builds/images.

By instead using subdomains, we can rely on the browser's built in
mechanisms for isolating domains to ensure that our builds don't
interact in unpredictable ways.

The new middleware in this commit maintains support for "hash" and
"branch" query string parameters.  When present, those parameters will
trigger a redirect to [commit-hash].[host-name].

The next middleware in the processing chain then grabs the commit hash
from the subdomain and stores it in the session, where dserve expects to
find it.