Prepend 'hash-' prefix to subdomain.
Previously, if you were to navigate to one hash and then append another
"?hash=xxx" query string, you'd end up adding another subdomain during
the redirect (e.g. [commit-hash-2].[commit-hash-1].[base-host-name]).

To avoid this, we need to strip the first commit hash off the front of
the host before redirecting to the second host.  That requires that we
be able to differentiate between the commit hash portion of the URL and
base host name.

We could achieve that by having a per-environment configuration
parameter specifying that base host name, but we don't currently have
per-environment configuration in dserve.  I don't think it's worth
adding that complexity for this purpose.  We could also look at the
format of the commit hash, but that could be unreliable if that format
changes (e.g. they change hash algorithms).

Instead, this commit uses a "hash-" prefix to identify the commit hash
segment of the host name when performing this stripping.