Add subdomain support to dserve (#62)
* Initial implementation of middleware for subdomains.

Current versions of dserve use a single domain for all commits and
branches.  This can allow cross-contamination of client-side resources
like cookies, localStorage and assets cached by service workers across
different builds/images.

By instead using subdomains, we can rely on the browser's built in
mechanisms for isolating domains to ensure that our builds don't
interact in unpredictable ways.

The new middleware in this commit maintains support for "hash" and
"branch" query string parameters.  When present, those parameters will
trigger a redirect to [commit-hash].[host-name].

The next middleware in the processing chain then grabs the commit hash
from the subdomain and stores it in the session, where dserve expects to
find it.

* Prepend 'hash-' prefix to subdomain.

Previously, if you were to navigate to one hash and then append another
"?hash=xxx" query string, you'd end up adding another subdomain during
the redirect (e.g. [commit-hash-2].[commit-hash-1].[base-host-name]).

To avoid this, we need to strip the first commit hash off the front of
the host before redirecting to the second host.  That requires that we
be able to differentiate between the commit hash portion of the URL and
base host name.

We could achieve that by having a per-environment configuration
parameter specifying that base host name, but we don't currently have
per-environment configuration in dserve.  I don't think it's worth
adding that complexity for this purpose.  We could also look at the
format of the commit hash, but that could be unreliable if that format
changes (e.g. they change hash algorithms).

Instead, this commit uses a "hash-" prefix to identify the commit hash
segment of the host name when performing this stripping.

* Use new middleware to support subdomains.

* Use redirect() method rather than manually setting status code and Location header.

* Use regex for subdomain filtering rather than splitting and iterating over each segment.

* Use CommitHash type rather than primitive string.

* Remove unused imports from middlewares.

* Use const and inline || operator rather than let and if/else.

* Remove unnecessary async declaration from middleware.

It looks like determineCommitHash has been unnecessarily marked async
for a while.  Perhaps the mapping of branch name to commit hash was
previously async.  That's no longer the case, so we can drop it now.

* Update middleware information in the README.