Recognize containers that were deleted out from under dserve
Remove the build dir on a successful build to prevent us from picking it up and thinking it's an in progress build.
Refresh local images on a successful build to prevent a race on state.localImages between the refresher and requests.
Move the pendingHash deletion after refreshing local images to prevent a race there as well.