Middleware: Pass non-branch query parameters on redirect
Resolves #79

When we remap a given URL from `calypso.live?branch=...` to the
"subdomain" URL we have been stripping away the entire query string when
redirecting the browser. This causes problems when we want to pass flags
to Calypso or otherwise send query args.

In this change we're appending the existing query string without the
branch name specified so that we can pass everything else through.

No query arg named `branch` will pass through though.