Fix: Loops ran too often due (#97)
* Fix: Loops ran too often due

Loops initiated in builder.ts were running as fast as possible because the constants that spec the delay were not initialized when the loops were set up. The constants were not initialized because of a cyclic dependency between builder and api were they were defined. When a cyclic dependency is present, the initialization order is undefined and exported variables may not yet have a value.

* remove logging, fix tests