Wait for a commit to appear in git before adding it to the build queue
Right now we blindly add hashes to the build queue and hope that we've caught the new commit before someone tries to kick off a build. With the new e2e tests though, this isn't really true. We kick off builds moments after the PR hits, especially for renovate builds.

Instead, if we don't have the commit in the local repo, refresh the branches once and then check again. If we still don't have it, bail.